requirements.txt
pandas
numpy
scikit-learn
pymysql
python-dotenv
